Add ongoing notifications; tighten up layout/icons a lot.
The weather/notifications row is now as dense as it sensibly can be.
E.g., icons are thresholded and cropped before scaling, for max
detail. Weather temps are variable-width.

The rest of the screen is used for ongoing notifications, most recent
first, but they will never be cleared. Music is specifically added as
if it were a notification - the real ongoing notification has no
tickerText. Annoyingly, neither does Navigation. Next job: can we
read text out of RemoteViews in such cases? Really want walking
navigation instructions on my watch. :-)