Unfinished: button to show Remember The Milk tasks
This will be on button A, if replay is disabled. You need to fill in the
first 2 nulls at the top of RMilk.java, and uncomment one piece of
getTasksText at a time... next step is to implement a proper UI flow
through this process (login & authorise, pick a list, get tasks from
that list).